fix for last PR https://github.com/Frogging-Family/nvidia-all/pull/325 (#326)
* PKGBUILD: add missing libs/patch

* typo

* fix patch

* change the file extension of the patches to .diff

* feat: add support for improved NVIDIA module parameters

* change `cat` command to `echo -e` in heradoc for modprobe conf

* fix: issue with last PR https://github.com/Frogging-Family/nvidia-all/pull/325

* fix: revert `_extradir` to orginal value

* egl-x11 remove from hard dependencies

+++ PKGBUILD
@@ -1909,7 +1909,7 @@ EOF
 
 nvidia-utils-tkg() {
   pkgdesc="NVIDIA driver utilities and libraries for 'nvidia-tkg'"
-  depends=('libglvnd' 'mesa' 'vulkan-icd-loader' 'egl-x11')
+  depends=('libglvnd' 'mesa' 'vulkan-icd-loader')
   if [ "$_eglgbm" = "external" ]; then
     depends+=('egl-gbm')
   fi
@@ -1918,6 +1918,9 @@ nvidia-utils-tkg() {
     if (( ${pkgver%%.*} >= 590 )); then
       depends+=('egl-wayland2')
     fi
+  fi
+  if (( ${pkgver%%.*} >= 590 )); then
+    depends+=('egl-x11')
   fi
   optdepends=('gtk2: nvidia-settings (GTK+ v2)'
               'gtk3: nvidia-settings (GTK+ v3)'
@@ -1986,7 +1989,7 @@ nvidia-utils-tkg() {
     elif [[ -e libnvidia-egl-xlib.so.1.0.3 ]]; then
       install -D -m755 "libnvidia-egl-xlib.so.1.0.3" "${pkgdir}/usr/lib/libnvidia-egl-xlib.so.1.0.3"
     fi
-    if [[ -e 20_nvidia_xlib.json ]]; then
+    if (( ${pkgver%%.*} <= 590 )) && [[ -e 20_nvidia_xlib.json ]]; then
       install -D -m644 "20_nvidia_xlib.json" "${pkgdir}/usr/share/egl/egl_external_platform.d/20_nvidia_xlib.json"
     fi
     if [[ -e libnvidia-egl-xcb.so.1 ]]; then
@@ -1998,7 +2001,7 @@ nvidia-utils-tkg() {
     elif [[ -e libnvidia-egl-xcb.so.1.0.3 ]]; then
       install -D -m755 "libnvidia-egl-xcb.so.1.0.3" "${pkgdir}/usr/lib/libnvidia-egl-xcb.so.1.0.3"
     fi
-    if [[ -e 20_nvidia_xcb.json ]]; then
+    if (( ${pkgver%%.*} <= 590 )) && [[ -e 20_nvidia_xcb.json ]]; then
       install -D -m644 "20_nvidia_xcb.json" "${pkgdir}/usr/share/egl/egl_external_platform.d/20_nvidia_xcb.json"
     fi
 
@@ -2280,8 +2283,8 @@ if [ "$_dkms" = "false" ] || [ "$_dkms" = "full" ]; then
       provides=('NVIDIA-MODULE')
 
       cd ${_srcbase}-${pkgver}
-      #_extradir="/usr/lib/modules/$(</usr/src/linux/version)/extramodules"
-      _extradir="/usr/lib/modules/$(</proc/sys/kernel/osrelease)/extramodules"
+      _extradir="/usr/lib/modules/$(</usr/src/linux/version)/extramodules"
+      #_extradir="/usr/lib/modules/$(</proc/sys/kernel/osrelease)/extramodules"
       install -Dt "${pkgdir}${_extradir}" -m644 kernel-open/*.ko
       find "${pkgdir}" -name '*.ko' -exec strip --strip-debug {} +
       find "${pkgdir}" -name '*.ko' -exec xz {} +
